elastic/solutions/academic-rag-app-with-rbac/requirements.txt
flask==2.3.3
elasticsearch==8.10.0
python-dotenv==1.0.0
boto3==1.28.57
gunicorn==21.2.0
requests==2.31.0
flask-login==0.6.2
